Rename CollgeMajor schema to Major
