+++ app/config/environment.example.js
@@ -6,5 +6,6 @@ export const environment = {
   minimumFetchInterval: 15,
   maxFailedAttempt: 3,
   baseURL: 'http://192.168.1.145:3000/api/v1',
-  sentryDSN: ''
+  sentryDSN: '',
+  apiKey: '',
 };
