+++ app/services/college_major_sync_service.js
@@ -1,4 +1,4 @@
-import CollegeMajor from '../models/CollegeMajor'
+import Major from '../models/Major'
 import {itemsPerPage} from '../constants/sync_data_constant'
 import CollegeMajorApi from '../api/college_major_api'
 
@@ -14,20 +14,20 @@ const collegeMajorSyncService = (() => {
   // private method
   function _handleSaveMajor(collegeMajors) {
     collegeMajors.map(collegeMajor => {
-      CollegeMajor.create(collegeMajor)
+      Major.create(collegeMajor)
     });
   }
 
   function _syncAndRemoveByPage(page, totalPage, prevCollegeMajors = []) {
     if (page > totalPage) {
-      CollegeMajor.deleteAll()
+      Major.deleteAll()
       _handleSaveMajor(prevCollegeMajors)
       return
     }
 
     new CollegeMajorApi().load(page, (res) => {
       const allPage = Math.ceil(res.pagy.count / itemsPerPage)
-      _syncAndRemoveByPage(page+1, allPage, [...prevCollegeMajors, ...res.college_majors])
+      _syncAndRemoveByPage(page+1, allPage, [...prevCollegeMajors, ...res.majors])
     })
   }
 })()
